refactor(fe): update common components
